requirements.txt
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
distlib==0.3.7
filelock==3.13.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-MySQLdb==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
pipenv==2023.11.15
platformdirs==4.0.0
python-dotenv==1.0.0
setuptools==69.0.2
virtualenv==20.24.7
Werkzeug==3.0.1
